@@ -1,10 +1,10 @@
 package com.example.komsic.cryptoconverter.data.db;
 
 
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
-import android.arch.persistence.room.TypeConverters;
-import android.support.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
+import androidx.room.TypeConverters;
+import androidx.annotation.NonNull;
 
 import com.example.komsic.cryptoconverter.data.db.util.BooleanConverter;
 
