@@ -1,10 +1,10 @@
 package com.example.komsic.cryptoconverter.di.application;
 
 import android.app.Application;
-import android.arch.persistence.db.SupportSQLiteDatabase;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
-import android.support.annotation.NonNull;
+import androidx.sqlite.db.SupportSQLiteDatabase;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
+import androidx.annotation.NonNull;
 
 import com.example.komsic.cryptoconverter.data.db.CurrencyCardDao;
 import com.example.komsic.cryptoconverter.data.db.CurrencyDb;
