@@ -1,8 +1,8 @@
 package com.example.komsic.cryptoconverter.viewmodel;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
 import com.example.komsic.cryptoconverter.data.CurrencyRepository;
 
