@@ -1,7 +1,7 @@
 package com.example.komsic.cryptoconverter.data;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
 import android.os.AsyncTask;
 
 import com.example.komsic.cryptoconverter.data.db.CurrencyCard;
