@@ -1,7 +1,7 @@
 package com.example.komsic.cryptoconverter.data.db;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
 import android.content.Context;
 import android.content.res.Resources;
 import android.os.AsyncTask;
