@@ -1,6 +1,6 @@
 package com.example.komsic.cryptoconverter.data.db.util;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 public class BooleanConverter {
 
